+++ instrumentation/runtime/anr/src/test/kotlin/com/splunk/rum/instrumentation/anr/internal/extractor/RumAnrAttributesExtractorTest.kt
@@ -16,7 +16,7 @@
 
 package com.splunk.rum.instrumentation.anr.internal.extractor
 
-import android.content.Context
+import android.app.Application
 import androidx.test.core.app.ApplicationProvider
 import com.splunk.rum.agent.common.otel.internal.GlobalRumConstants
 import io.opentelemetry.api.common.Attributes
@@ -33,8 +33,8 @@ class RumAnrAttributesExtractorTest {
 
     @Before
     fun setUp() {
-        val context: Context = ApplicationProvider.getApplicationContext()
-        extractor = RumAnrAttributesExtractor(context)
+        val application: Application = ApplicationProvider.getApplicationContext()
+        extractor = RumAnrAttributesExtractor(application)
     }
 
     @Test